+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Revision history for small-bytearray-builder
 
-## 0.2.2.0 -- 2019-??-??
+## 0.3.0.0 -- 2019-??-??
 
 * Introduce `consLensBE32` for efficient serialization of wire protocols
   that require prefixing a payload with its length.
 * Add `int64BE` as a convenience.
 * Add little-endian encoding functions for `Word16`, `Word32`, and `Word64`.
+* Add `flush`, `copy`, and `insert` for better control when
+  converting byte sequences to builders.
 
 ## 0.2.1.0 -- 2019-09-05